refactor: migrate from neodbmodel to dbmodel and clean up unused files
