+++ src/qdash/api/main.py
@@ -8,7 +8,6 @@ from qdash.api.routers import (
     chip,
     execution,
     file,
-    fridges,
     menu,
     settings,
     task,
@@ -65,7 +64,6 @@ app.add_middleware(
 app.include_router(calibration.router, tags=["calibration"])
 app.include_router(menu.router, tags=["menu"])
 app.include_router(settings.router, tags=["settings"])
-app.include_router(fridges.router, tags=["fridges"])
 app.include_router(execution.router, tags=["execution"])
 app.include_router(chip.router, tags=["chip"])
 app.include_router(file.router, tags=["file"])
