+++ src/qdash/api/schemas/service.py
@@ -2,14 +2,15 @@ from pydantic import BaseModel
 
 
 class ServiceStatusResponse(BaseModel):
-    """
-    Represents the response for the service status.
+    """Represents the response for the service status.
 
-    Attributes:
+    Attributes
+    ----------
         status (str): The status of the service.
         backend (str): The backend used by the service.
         qubit_num (int): The number of qubits.
         qubit_index (str): The index of the qubit.
+
     """
 
     status: str
@@ -32,14 +33,15 @@ class ServiceStatusResponse(BaseModel):
 
 
 class ServiceStatusRequest(BaseModel):
-    """
-    Represents a request to get the status of a service.
+    """Represents a request to get the status of a service.
 
-    Attributes:
+    Attributes
+    ----------
         status (str): The status of the service.
         backend (str): The backend of the service.
         qubit_num (int): The number of qubits in the service.
         qubit_index (str): The index of the qubit in the service.
+
     """
 
     status: str
@@ -62,11 +64,12 @@ class ServiceStatusRequest(BaseModel):
 
 
 class RestartServiceResponse(BaseModel):
-    """
-    Represents the response for restarting a service.
+    """Represents the response for restarting a service.
 
-    Attributes:
+    Attributes
+    ----------
         message (str): The message indicating the success of the restart.
+
     """
 
     message: str = "QMT server restarted"
@@ -83,11 +86,12 @@ class RestartServiceResponse(BaseModel):
 
 
 class UpdateTopologyInfoResponse(BaseModel):
-    """
-    Represents the response for updating the topology information.
+    """Represents the response for updating the topology information.
 
-    Attributes:
+    Attributes
+    ----------
         message (str): The message indicating the success of the update.
+
     """
 
     message: str = "Topology information updated"
@@ -104,11 +108,12 @@ class UpdateTopologyInfoResponse(BaseModel):
 
 
 class UpdateWiringInfoResponse(BaseModel):
-    """
-    Represents the response for updating the wiring information.
+    """Represents the response for updating the wiring information.
 
-    Attributes:
+    Attributes
+    ----------
         message (str): The message indicating the success of the update.
+
     """
 
     message: str = "Wiring information updated"
