+++ src/qdash/dbmodel/chip.py
