+++ src/qdash/dbmodel/fridge.py
