+++ src/qdash/dbmodel/parameter.py
