+++ src/qdash/workflow/plugin/bluefors/blocks.py
@@ -1,8 +1,8 @@
-from prefect.blocks.core import Block
+# from prefect.blocks.core import Block
 
 
-class BlueforsBlock(Block):
-    bluefors_server_url: str = ""
-    mongo_url: str = ""
-    mongo_database: str = ""
-    mongo_collection: str = ""
+# class BlueforsBlock(Block):
+#     bluefors_server_url: str = ""
+#     mongo_url: str = ""
+#     mongo_database: str = ""
+#     mongo_collection: str = ""
